# Swagger/api_mod.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "API",
    "description": "API for Portal System in AEM",
    "termsOfService": "None"
  },
  "host": "otvcrm01:84",
  "basePath": "/",
  "schemes": [ "http", "https" ],
  "consumes": [ "application/json" ],
  "produces": [ "application/json" ],
  "paths": {
...
    "/api/FundingOpportunityInternalUsers": {
      "get": {
        "tags": [ "FundingOpportunityInternalUsers" ],
        "operationId": "ApiFundingOpportunityInternalUsersGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunityInternalUser" }
            }
          }
        }
      },
      "post": {
        "tags": [ "FundingOpportunityInternalUsers" ],
        "operationId": "ApiFundingOpportunityInternalUsersPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "fundingOpportunityInternalUser",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/FundingOpportunityInternalUser" }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FundingOpportunityInternalUser" }
          }
        }
      }
    },
    "/api/FundingOpportunityInternalUsers/{id}": {
      "get": {
        "tags": [ "FundingOpportunityInternalUsers" ],
        "operationId": "ApiFundingOpportunityInternalUsersByIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FundingOpportunityInternalUser" }
          }
        }
      },
      "put": {
        "tags": [ "FundingOpportunityInternalUsers" ],
        "operationId": "ApiFundingOpportunityInternalUsersByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "fundingOpportunityInternalUser",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/FundingOpportunityInternalUser" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "FundingOpportunityInternalUsers" ],
        "operationId": "ApiFundingOpportunityInternalUsersByIdDelete",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FundingOpportunityInternalUser" }
          }
        }
      }
    },
    "/api/FundingOpportunityObjectives": {
      "get": {
        "tags": [ "FundingOpportunityObjectives" ],
        "operationId": "ApiFundingOpportunityObjectivesGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/FundingOpportunityObjective" }
            }
          }
        }
      },
      "post": {
        "tags": [ "FundingOpportunityObjectives" ],
        "operationId": "ApiFundingOpportunityObjectivesPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "fundingOpportunityObjective",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/FundingOpportunityObjective" }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FundingOpportunityObjective" }
          }
        }
      }
    },
    "/api/FundingOpportunityObjectives/{id}": {
      "get": {
        "tags": [ "FundingOpportunityObjectives" ],
        "operationId": "ApiFundingOpportunityObjectivesByIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FundingOpportunityObjective" }
          }
        }
      },
      "put": {
        "tags": [ "FundingOpportunityObjectives" ],
        "operationId": "ApiFundingOpportunityObjectivesByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "fundingOpportunityObjective",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/FundingOpportunityObjective" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "FundingOpportunityObjectives" ],
        "operationId": "ApiFundingOpportunityObjectivesByIdDelete",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/FundingOpportunityObjective" }
          }
        }
      }
    },
...
    "/api/InternalUsers/{id}": {
      "get": {
        "tags": [ "InternalUsers" ],
        "operationId": "ApiInternalUsersByIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/InternalUser" }
          }
        }
      },
      "put": {
        "tags": [ "InternalUsers" ],
        "operationId": "ApiInternalUsersByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "internalUser",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/InternalUser" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "InternalUsers" ],
        "operationId": "ApiInternalUsersByIdDelete",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/InternalUser" }
          }
        }
      }
    },
...
    "/api/Objectives/{id}": {
      "get": {
        "tags": [ "Objectives" ],
        "operationId": "ApiObjectivesByIdGet",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Objective" }
          }
        }
      },
      "put": {
        "tags": [ "Objectives" ],
        "operationId": "ApiObjectivesByIdPut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "objective",
            "in": "body",
            "required": false,
            "schema": { "$ref": "#/definitions/Objective" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "delete": {
        "tags": [ "Objectives" ],
        "operationId": "ApiObjectivesByIdDelete",
        "consumes": [],
        "produces": [ "application/json" ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": { "$ref": "#/definitions/Objective" }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "InternalUser": {
      "type": "object",
      "properties": {
        "internalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "userName": { "type": "string" },
        "firstName": { "type": "string" },
        "lastName": { "type": "string" },
        "email": { "type": "string" },
        "language": { "type": "string" },
        "isPortalAdministrator": { "type": "boolean" },
        "isFundingOpportunityAdministrator": { "type": "boolean" },
        "isWorkloadManager": { "type": "boolean" },
        "isSubmissionReviewer": { "type": "boolean" },
        "fullName": { "type": "string" },
        "roles": { "type": "string" },
        "fundingProgramInternalUsers": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingProgramInternalUser" }
        },
        "fundingOpportunityInternalUsers": {
          "type": "array",
          "items": { "$ref": "#/definitions/FundingOpportunityInternalUser" }
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
    "FundingProgramInternalUser": {
      "type": "object",
      "properties": {
        "fundingProgramInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingProgramId": {
          "format": "uuid",
          "type": "string"
        },
        "internalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "FundingOpportunityInternalUser": {
      "type": "object",
      "properties": {
        "fundingOpportunityInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "internalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "hasFullAccess": { "type": "boolean" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
...
    "FundingOpportunityObjective": {
      "required": [ "fundingOpportunityId", "objectiveId" ],
      "type": "object",
      "properties": {
        "fundingOpportunityObjectiveId": {
          "format": "uuid",
          "type": "string"
        },
        "fundingOpportunityId": {
          "format": "uuid",
          "type": "string"
        },
        "objectiveId": {
          "format": "uuid",
          "type": "string"
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
    "Objective": {
      "required": [ "descriptionE", "descriptionF" ],
      "type": "object",
      "properties": {
        "objectiveId": {
          "format": "uuid",
          "type": "string"
        },
        "gcimsAttributeID": { "type": "string" },
        "lang": { "type": "string" },
        "description": { "type": "string" },
        "descriptionE": { "type": "string" },
        "descriptionF": { "type": "string" },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        },
        "updatedOn": {
          "format": "date-time",
          "type": "string"
        },
        "createdByInternalUserId": {
          "format": "uuid",
          "type": "string"
        },
        "updatedByInternalUserId": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {}
}